+++ CHANGELOG.rst
@@ -121,11 +121,13 @@ Fixed
     - Nested paths set by assay plugin reported as orphaned (#2207)
 - **Landingzones**
     - Missing investigation setup in certain ``TestProjectZoneView`` tests (#2290)
+    - Missing ``HTTPRefererMixin`` inheritance in views (#206)
 - **Samplesheets**
     - Details card iRODS stats badge fails with guest role (#2243, #2271)
     - Sheet Operations dropdown disabled for guest user (#2244)
     - Assay table title row content wrapping (#2267)
     - Vue app modal close button alignment (#2304)
+    - Missing ``HTTPRefererMixin`` inheritance in views (#206)
 - **Taskflowbackend**
     - Incorrect user in landing zone move sheet cache update timeline event (#2152)